+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.4.0-beta1 (2017-XX-XX)
 
+ * Add the contao.web_dir parameter (see contao/installation-bundle#40).
  * Look up the form class and allow to choose the type (see contao/core#8527). 
  * Auto-select the active page in the quick navigation/link module (see contao/core#8587).
  * Add the contao.root_dir parameter (see #662).